LNURL-withdraw: move remaining tests to sdk-common (E2E tests, but only cover first portion of the flow, without payment)
